config service is now its own thing
