+++ src/services/train_run_template/__init__.py
@@ -1,0 +1,3 @@
+from services.train_run_template.TrainRunTemplateConfig import TrainRunTemplateConfig
+
+__all__ = ["TrainRunTemplateConfig"]
